services/payment-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
opentelemetry-sdk==1.29.0
opentelemetry-api==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-http==1.29.0
azure-monitor-opentelemetry==1.6.4
httpx==0.27.2
python-dotenv==1.0.1
